Formatting result messages, adding a random query, and updating build files to include yelp api key

+++ model/yelp.go
@@ -7,7 +7,7 @@ type SearchResponse struct {
 }
 
 type Business struct {
-	Rating       float32     `json:"rating"`
+	Rating       float64     `json:"rating"`
 	Price        string      `json:"price"`
 	Phone        string      `json:"phone"`
 	ID           string      `json:"id"`
